project-04/requirements.txt
sagemaker-pyspark==1.4.1